Bounded page cache (#25)
* Bounded CLOCK-Pro page cache replacing unbounded HashMap

Replace the unbounded RwLock<HashMap> read cache in PagedNodeStorage with a
bounded quick_cache::sync::Cache (CLOCK-Pro eviction). The cache is capped at
16,384 pages (~64 MB) by default, providing scan resistance so range scans
over cold leaves don't evict hot root/internal nodes.

Also adds bench_epoch benchmark and reduces concurrent_stress test rounds
from 50 to 5 for reasonable debug-build runtime.

* Clarify frequency vs recency distinction in CLOCK-Pro docs

* Fixed redundant locking pattern in epoch reclamation loop

* Reduce lock contention in reclaim() and remove unnecessary clones in txn commit

* Hoist Vec allocations above transaction retry loop

* Rewrote reclamation loops in a more idiomatic/functional style

* Hoist as usize cast out of merge_into loop bodies

* Added writer contention bench tests

+++ src/bplustree/node_view.rs
@@ -411,11 +411,11 @@ impl NodeView {
                     page: other_page, ..
                 },
             ) => {
-                let other_key_count = other_page.key_count();
+                let other_key_count = other_page.key_count() as usize;
                 let mut scratch = ScratchBuf::new();
                 for i in 0..other_key_count {
-                    let key = other_page.get_key_at(i as usize, &mut scratch)?;
-                    let child_ptr = other_page.read_child_at(i as usize + 1)?;
+                    let key = other_page.get_key_at(i, &mut scratch)?;
+                    let child_ptr = other_page.read_child_at(i + 1)?;
                     self_page.append(key, child_ptr)?;
                 }
                 Ok(())
@@ -428,10 +428,10 @@ impl NodeView {
                     page: other_page, ..
                 },
             ) => {
-                let other_key_count = other_page.key_count();
+                let other_key_count = other_page.key_count() as usize;
                 let mut scratch = ScratchBuf::new();
                 for i in 0..other_key_count {
-                    let (k, v) = other_page.get_kv_at(i as usize, &mut scratch)?;
+                    let (k, v) = other_page.get_kv_at(i, &mut scratch)?;
                     self_page.append(k, v)?;
                 }
                 Ok(())
